pyproject.toml
[project]
name = "orkestr8-sdk"
version="0.0.1"
requires-python=">=3.8"
authors=[{name="Antsthebul"}]
description="SDK to easily interact with Orkestr8 ML entities"
dependencies=[
    "python-dotenv",
    "boto3"
]

[project.scripts]
orkestr8="orkestr8:run"